requirements.txt
discord.py
psutil
asyncio
mcstatus